build: move dev server generation into deploy.Build function
