+++ testing/testing.go
@@ -127,11 +127,6 @@ func Run(ctx context.Context, opts *RunnerOpts) error {
 		}
 	}
 
-	err = os.WriteFile(path.Join(opts.Dir, ".build/server.js"), []byte(functionsServerEntry), os.ModePerm)
-	if err != nil {
-		return err
-	}
-
 	runtimePort, err := util.GetFreePort()
 	if err != nil {
 		return err
@@ -498,40 +493,3 @@ func toLambdaFunctionURLRequest(r *http.Request) (lambdaevents.LambdaFunctionURL
 		IsBase64Encoded: false,
 	}, nil
 }
-
-const functionsServerEntry = `
-const { createServer } = require("node:http");
-const { handler } = require("./functions/main.js");
-
-const server = createServer(async (req, res) => {
-	try {
-		const u = new URL(req.url, "http://" + req.headers.host);
-		if (req.method === "GET" && u.pathname === "/_health") {
-			res.statusCode = 200;
-			res.end();
-			return;
-		}
-
-		const buffers = [];
-		for await (const chunk of req) {
-			buffers.push(chunk);
-		}
-		const data = Buffer.concat(buffers).toString();
-		const json = JSON.parse(data);
-
-		const rpcResponse = await handler(json, {});
-		res.statusCode = 200;
-		res.setHeader('Content-Type', 'application/json');
-		res.write(JSON.stringify(rpcResponse));
-		res.end();
-	} catch (err) {
-		res.status = 400;
-		res.write(err.message);
-	}
-
-	res.end();
-});
-
-const port = (process.env.PORT && parseInt(process.env.PORT, 10)) || 3001;
-server.listen(port);
-`
